codegen(python): preserve exclude-none for plain models + move coverage fixtures out of the demo
- Restore the historical wire format for plain Pydantic request models:
  `model_dump_json(by_alias=True, exclude_none=True)` for non-partial
  models so unset optional fields stay absent on the wire.
  `ReflectapiPartialModel` still uses no exclude — explicit `None`
  encodes the protocol's "field present, value cleared" state and
  must reach the wire. Same fix applied in `client.py` and `streaming.py`.

- Move codegen coverage fixtures (PyKeywordFields, recursive types,
  Pydantic-reserved field names, etc.) out of the public demo schema
  into the integration test at `reflectapi-demo/tests/codegen_coverage.rs`.
  The demo crate's `builder()` now only registers pet-themed routes.
  The integration test builds a private schema, runs Python codegen to
  `target/codegen-coverage-client/`, and CI imports the result so
  strict `model_rebuild()` still catches dangling-reference bugs.

- Add four runtime tests pinning the partial-vs-plain wire-format
  contract.
